@@ -21,33 +21,30 @@
   "scripts": {
     "test": "npm run lint && npm run check && npm run testonly",
     "test:ci": "npm run lint && npm run check && npm run testonly -- --coverage",
-    "lint": "eslint src",
-    "check": "flow check --max-warnings 0",
-    "build": "babel src --ignore src/__tests__ --out-dir dist/ ; cp src/index.js dist/index.js.flow ; cp src/index.d.ts dist/",
-    "watch": "babel resources/watch.js | node",
+    "lint": "eslint src --ext .ts",
+    "check": "tsc --noEmit",
+    "build": "tsc",
+    "watch": "tsc --watch",
     "testonly": "jest src",
     "prerelease": ". ./resources/prepublish.sh",
     "release": "changeset publish"
   },
   "files": [
     "index.js",
-    "index.js.flow",
     "index.d.ts",
     "README.md",
     "LICENSE",
     "PATENTS"
   ],
   "devDependencies": {
-    "@babel/cli": "7.7.0",
-    "@babel/core": "7.7.2",
-    "@babel/node": "7.7.0",
-    "@babel/preset-env": "7.7.1",
-    "@babel/preset-flow": "7.0.0",
     "@changesets/cli": "^2.22.0",
-    "babel-eslint": "10.0.3",
-    "eslint": "6.6.0",
-    "flow-bin": "0.112.0",
-    "jest": "24.9.0",
+    "@typescript-eslint/eslint-plugin": "^5.18.0",
+    "@typescript-eslint/parser": "^5.18.0",
+    "@types/jest": "27.4.1",
+    "eslint": "8.13.0",
+    "jest": "27.5.1",
+    "ts-jest": "27.1.4",
+    "typescript": "4.6.3",
     "sane": "4.1.0"
   },
   "publishConfig": {
