@@ -3,17 +3,15 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-const DataLoader = require('..');
+import DataLoader from '..';
 
 describe('Provides descriptive error messages for API abuse', () => {
 
   it('Loader creation requires a function', () => {
     expect(() => {
-      // $FlowExpectError
+      // @ts-expect-error
       new DataLoader(); // eslint-disable-line no-new
     }).toThrow(
       'DataLoader must be constructed with a function which accepts ' +
@@ -21,7 +19,7 @@ describe('Provides descriptive error messages for API abuse', () => {
     );
 
     expect(() => {
-      // $FlowExpectError
+      // @ts-expect-error
       new DataLoader({}); // eslint-disable-line no-new
     }).toThrow(
       'DataLoader must be constructed with a function which accepts ' +
@@ -33,7 +31,7 @@ describe('Provides descriptive error messages for API abuse', () => {
     const idLoader = new DataLoader<number, number>(async keys => keys);
 
     expect(() => {
-      // $FlowExpectError
+      // @ts-expect-error
       idLoader.load();
     }).toThrow(
       'The loader.load() function must be called with a value, ' +
@@ -41,7 +39,6 @@ describe('Provides descriptive error messages for API abuse', () => {
     );
 
     expect(() => {
-      // $FlowExpectError
       idLoader.load(null);
     }).toThrow(
       'The loader.load() function must be called with a value, ' +
@@ -58,7 +55,7 @@ describe('Provides descriptive error messages for API abuse', () => {
     const idLoader = new DataLoader<number, number>(async keys => keys);
 
     expect(() => {
-      // $FlowExpectError
+      // @ts-expect-error
       idLoader.loadMany();
     }).toThrow(
       'The loader.loadMany() function must be called with Array<key> ' +
@@ -66,7 +63,7 @@ describe('Provides descriptive error messages for API abuse', () => {
     );
 
     expect(() => {
-      // $FlowExpectError
+      // @ts-expect-error
       idLoader.loadMany(1, 2, 3);
     }).toThrow(
       'The loader.loadMany() function must be called with Array<key> ' +
@@ -80,7 +77,6 @@ describe('Provides descriptive error messages for API abuse', () => {
   });
 
   it('Batch function must return a Promise, not null', async () => {
-    // $FlowExpectError
     const badLoader = new DataLoader<number, number>(() => null);
 
     let caughtError;
@@ -90,7 +86,7 @@ describe('Provides descriptive error messages for API abuse', () => {
       caughtError = error;
     }
     expect(caughtError).toBeInstanceOf(Error);
-    expect((caughtError: any).message).toBe(
+    expect((caughtError as any).message).toBe(
       'DataLoader must be constructed with a function which accepts ' +
       'Array<key> and returns Promise<Array<value>>, but the function did ' +
       'not return a Promise: null.'
@@ -99,7 +95,7 @@ describe('Provides descriptive error messages for API abuse', () => {
 
   it('Batch function must return a Promise, not a value', async () => {
     // Note: this is returning the keys directly, rather than a promise to keys.
-    // $FlowExpectError
+    // @ts-expect-error
     const badLoader = new DataLoader<number, number>(keys => keys);
 
     let caughtError;
@@ -109,7 +105,7 @@ describe('Provides descriptive error messages for API abuse', () => {
       caughtError = error;
     }
     expect(caughtError).toBeInstanceOf(Error);
-    expect((caughtError: any).message).toBe(
+    expect((caughtError as any).message).toBe(
       'DataLoader must be constructed with a function which accepts ' +
       'Array<key> and returns Promise<Array<value>>, but the function did ' +
       'not return a Promise: 1.'
@@ -128,7 +124,7 @@ describe('Provides descriptive error messages for API abuse', () => {
       caughtError = error;
     }
     expect(caughtError).toBeInstanceOf(Error);
-    expect((caughtError: any).message).toBe(
+    expect((caughtError as any).message).toBe(
       'DataLoader must be constructed with a function which accepts ' +
       'Array<key> and returns Promise<Array<value>>, but the function did ' +
       'not return a Promise of an Array: null.'
@@ -146,7 +142,7 @@ describe('Provides descriptive error messages for API abuse', () => {
       caughtError = error;
     }
     expect(caughtError).toBeInstanceOf(Error);
-    expect((caughtError: any).message).toBe(
+    expect((caughtError as any).message).toBe(
       'DataLoader must be constructed with a function which accepts ' +
       'Array<key> and returns Promise<Array<value>>, but the function did ' +
       'not return a Promise of an Array of the same length as the Array ' +
@@ -162,9 +158,9 @@ describe('Provides descriptive error messages for API abuse', () => {
     }
 
     expect(() => {
-      // $FlowExpectError
       const incompleteMap = new IncompleteMap();
       const options = { cacheMap: incompleteMap };
+      // @ts-expect-error
       new DataLoader(async keys => keys, options); // eslint-disable-line no-new
     }).toThrow(
       'Custom cacheMap missing methods: set, delete, clear'
@@ -173,7 +169,6 @@ describe('Provides descriptive error messages for API abuse', () => {
 
   it('Requires a number for maxBatchSize', () => {
     expect(() =>
-      // $FlowExpectError
       new DataLoader(async keys => keys, { maxBatchSize: null })
     ).toThrow('maxBatchSize must be a positive number: null');
   });
@@ -193,7 +188,6 @@ describe('Provides descriptive error messages for API abuse', () => {
 
   it('Requires a function for batchScheduleFn', () => {
     expect(() =>
-      // $FlowExpectError
       new DataLoader(async keys => keys, { batchScheduleFn: null })
     ).toThrow('batchScheduleFn must be a function: null');
   });
