@@ -3,18 +3,15 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
- *
- * @flow
  */
 
-const DataLoader = require('..');
+import DataLoader from '..';
 
 describe('Unhandled rejections', () => {
   it('Not catching a primed error is an unhandled rejection', async () => {
     let hadUnhandledRejection = false;
-    // Override Jest's unhandled detection
-    global.jasmine.process.removeAllListeners('unhandledRejection');
-    global.jasmine.process.on('unhandledRejection', () => {
+    process.removeAllListeners('unhandledRejection');
+    process.addListener('unhandledRejection', () => {
       hadUnhandledRejection = true;
     });
 
